### 2020-05-22 #2206 @whedon
[PRE REVIEW]: BrkRaw: A comprehensive tool for accessing raw Bruker Biospin MRI data
**Submitting author:** @dvm-shlee (<a href="http://orcid.org/0000-0001-5292-0747">Sung-Ho Lee</a>)
**Repository:** <a href="https://github.com/brkraw/bruker" target ="_blank">https://github.com/brkraw/bruker</a>
**Version:** 0.3.3
