### 2020-05-22 #2206 @arfon
👋 @richardjgowers - would you be able to edit this for JOSS? I realize it's not really in your field but our MRI specialist is completely unavailable to edit currently.
