+++ brkraw/__init__.py
@@ -1,6 +1,6 @@
 from .lib import *
 
-__version__ = '0.3.3rc2'
+__version__ = '0.3.3rc3'
 __all__ = ['BrukerLoader', '__version__']
 
 
