### 2020-06-01 372d3b7
update nifti convet function and documents
