### 2020-06-01 a802d8e
[bug patch] (rc5) backup.py Backup validation generates unexpected error
