+++ brkraw/lib/backup.py
@@ -626,21 +626,12 @@ class BackupCacheHandler:
 
                         # Backup validation
                         if not os.path.exists(tmp_path):  # Check if the file is generated
-                            print_internal_error(fobj)
                             error = ArchiveFailedError(raw_path)
                             self.logging(error.message, 'backup')
                             raise error
                         else:
                             try:
                                 os.rename(tmp_path, arc_path)
-                            except OSError:
+                            except:
                                 print_internal_error(fobj)
-                                error = RenameFailedError(tmp_path, arc_path)
-                                self.logging(error.message, 'backup')
-                                raise error
-                            else:
-                                print_internal_error(fobj)
-                                error = UnexpectedError
-                                self.logging(error.message, 'backup')
-
-                                raise error
+                                raise UnexpectedError
