+++ setup.py
@@ -25,7 +25,7 @@ setup(name='bruker',
       url=__url__,
       license='GNLv3',
       packages=find_packages(),
-      install_requires=['shleeh>=0.0.4',
+      install_requires=['shleeh>=0.0.7',
                         'nibabel>=3.0.2',
                         'SimpleITK>=1.2.4',
                         'numpy>=1.18.0',
