### 2020-06-02 0c2bb6f
[bug pathes] Issue #2,3,5,6
