### 2020-06-02 4c79824
[bug pathes] Issue #2,3,5,6
