### 2020-06-02 45d0701
[bug pathes] Issue #2,3,5,6
