+++ brkraw/lib/loader.py
@@ -385,15 +385,15 @@ class BrukerLoader():
 
     def _inspect_ids(self, scan_id, reco_id):
         if scan_id not in self._avail.keys():
-            print(f'Invalid Scan ID.\n'
-                  f'Your input: {scan_id}\n'
-                  f'Available Scan IDs: {list(self._avail.keys())}')
+            print(f'[Error] Invalid Scan ID.\n'
+                  f'  - Your input: {scan_id}\n'
+                  f'  - Available Scan IDs: {list(self._avail.keys())}')
             raise ValueError
         else:
             if reco_id not in self._avail[scan_id]:
-                print(f'Invalid Reco ID.\n'
-                      f'Your input: {reco_id}\n'
-                      f'Available Reco IDs: {self._avail[scan_id]}')
+                print(f'[Error] Invalid Reco ID.\n'
+                      f'  - Your input: {reco_id}\n'
+                      f'  - Available Reco IDs: {self._avail[scan_id]}')
                 raise ValueError
 
     def save_nifti(self, scan_id, reco_id, filename, dir='./', ext='nii.gz',
