### 2020-06-02 0d0f854
[bug pathes] Issue #2,#3,#5,#6
