### 2020-06-02 0a9295b
[add feature] Issue #7- brkraw CLI now take --ignore_slope option
