+++ brkraw/scripts/brkraw.py
@@ -49,6 +49,7 @@ def main():
 
     gui.add_argument("-i", "--input", help=input_str, type=str, default=None)
     gui.add_argument("-o", "--output", help=output_dir_str, type=str, default=None)
+    gui.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
 
     nii.add_argument("input", help=input_str, type=str)
     nii.add_argument("-b", "--bids", help=bids_opt, action='store_true')
@@ -56,10 +57,12 @@ def main():
     nii.add_argument("-r", "--recoid", help="RECO ID (default=1), option to specify a particular reco to convert",
                      type=int, default=1)
     nii.add_argument("-s", "--scanid", help="Scan ID, option to specify a particular scan to convert.", type=str)
+    nii.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
 
     niiall.add_argument("input", help=input_dir_str, type=str)
     niiall.add_argument("-o", "--output", help=output_dir_str, type=str)
     niiall.add_argument("-b", "--bids", help=bids_opt, action='store_true')
+    niiall.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
 
     bids_helper.add_argument("input", help=input_dir_str, type=str)
     bids_helper.add_argument("output", help="output BIDS datasheet filename (.xlsx)", type=str)
@@ -70,6 +73,7 @@ def main():
     bids_convert.add_argument("datasheet", help="input BIDS datahseet filename", type=str)
     bids_convert.add_argument("-j", "--json", help="input JSON syntax template filename", type=str, default=False)
     bids_convert.add_argument("-o", "--output", help=output_dir_str, type=str, default=False)
+    bids_convert.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
 
     args = parser.parse_args()
 
@@ -93,6 +97,8 @@ def main():
         root = MainWindow()
         if ipath != None:
             root._path = ipath
+            if args.ignore_slope is True:
+                root._ignore_slope = True
             root._extend_layout()
             root._load_dataset()
         if opath != None:
@@ -106,6 +112,10 @@ def main():
         scan_id = args.scanid
         reco_id = args.recoid
         study = BrukerLoader(path)
+        if args.ignore_slope:
+            slope = None
+        else:
+            slope = False
         if args.output:
             output = args.output
         else:
@@ -115,7 +125,7 @@ def main():
             try:
                 scan_id = int(scan_id)
                 reco_id = int(reco_id)
-                study.save_as(scan_id, reco_id, output_fname)
+                study.save_as(scan_id, reco_id, output_fname, slope=slope)
                 if args.bids:
                     study.save_json(scan_id, reco_id, output_fname)
                 print('NifTi file is generated... [{}]'.format(output_fname))
@@ -126,7 +136,7 @@ def main():
                 for reco_id in recos:
                     output_fname = '{}-{}-{}'.format(output, str(scan_id).zfill(2), reco_id)
                     try:
-                        study.save_as(scan_id, reco_id, output_fname)
+                        study.save_as(scan_id, reco_id, output_fname, slope=slope)
                         if args.bids:
                             study.save_json(scan_id, reco_id, output_fname)
                         print('NifTi file is genetared... [{}]'.format(output_fname))
@@ -135,6 +145,10 @@ def main():
 
     elif args.function == 'tonii_all':
         path = args.input
+        if args.ignore_slope:
+            slope = None
+        else:
+            slope = False
         from os.path import join as opj, isdir, isfile
         list_of_raw = sorted([d for d in os.listdir(path) if isdir(opj(path, d)) \
[…]
