### 2020-06-02 d37c7ab
[add feature] Issue #8- bdata generation
