### 2020-06-04 18185d3
[resolve issue] #3 and #7
