+++ brkraw/lib/loader.py
@@ -139,7 +139,7 @@ class BrukerLoader():
             data_off = data_off[0] if is_all_element_same(data_off) else data_off
         return data_slp, data_off
 
-    def get_dataobj(self, scan_id, reco_id, slope=True):
+    def get_dataobj(self, scan_id, reco_id, slope=True, offset=True):
         """ Return dataobj that has 3D(spatial) + extra frame
         Args:
             scan_id: scan id
@@ -155,8 +155,10 @@ class BrukerLoader():
         dataobj = self._get_dataobj(scan_id, reco_id)
         group_id    = fg_info['group_id']
 
+        data_slp, data_off = self._get_dataslp(visu_pars)
+
         if slope:
-            data_slp, data_off = self._get_dataslp(visu_pars)
+            # This option apply the slope to data array directly instead of header
             f = fg_info['frame_size']
             if isinstance(data_slp, list):
                 if f != len(data_slp):
@@ -172,9 +174,30 @@ class BrukerLoader():
                     else:
                         raise UnexpectedError(message='Unexpected frame shape on DTI image;'
                                                       f'{ISSUE_REPORT}')
-                dataobj = (_dataobj * data_slp + data_off).T
+                dataobj = (_dataobj * data_slp).T
             else:
-                dataobj = dataobj * data_slp + data_off
+                dataobj = dataobj * data_slp
+
+        if offset:
+            # This option apply the offset to data array directly instead of header
+            f = fg_info['frame_size']
+            if isinstance(data_off, list):
+                if f != len(data_off):
+                    raise UnexpectedError(message='data_off mismatch;'
+                                                  f'{ISSUE_REPORT}')
+                else:
+                    if dim == 2:
+                        x, y = matrix_size[:2]
+                        _dataobj = dataobj.reshape([f, x * y]).T
+                    elif dim == 3:
+                        x, y, z = matrix_size[:3]
+                        _dataobj = dataobj.reshape([f, x * y * z]).T
+                    else:
+                        raise UnexpectedError(message='Unexpected frame shape on DTI image;'
+                                                      f'{ISSUE_REPORT}')
+                dataobj = (_dataobj + data_off).T
+            else:
+                dataobj = dataobj + data_off
 
         dataobj = dataobj.reshape(matrix_size[::-1]).T
 
@@ -242,7 +265,7 @@ class BrukerLoader():
 
     # methods to dump data into file object
     ## - NifTi1
-    def get_niftiobj(self, scan_id, reco_id, crop=None, slope=False):
+    def get_niftiobj(self, scan_id, reco_id, crop=None, slope=False, offset=False):
         """ return nibabel nifti object
         Args:
             scan_id:
@@ -276,7 +299,7 @@ class BrukerLoader():
                 end = start + num_slices_each_pack[spack_idx]
                 seg_imgobj = imgobj[..., start:end]
                 niiobj = Nifti1Image(seg_imgobj, np.round(affine[spack_idx], decimals=3))
-                niiobj = self._set_nifti_header(niiobj, visu_pars, method, slope=slope)
+                niiobj = self._set_nifti_header(niiobj, visu_pars, method, slope=slope, offset=offset)
                 parser.append(niiobj)
             return parser
         affine = np.round(affine, decimals=3)
@@ -299,7 +322,7 @@ class BrukerLoader():
                         niiobj_ = Nifti1Image(imgobj_[..., crop[0]:crop[1]], affine)
                 else:
                     niiobj_ = Nifti1Image(imgobj_, affine)
[…]
