+++ brkraw/scripts/brkraw.py
@@ -49,20 +49,27 @@ def main():
 
     gui.add_argument("-i", "--input", help=input_str, type=str, default=None)
     gui.add_argument("-o", "--output", help=output_dir_str, type=str, default=None)
-    gui.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
+    gui.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
+    gui.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
+    gui.add_argument("--ignore-rescale", help='remove slope and offset value from header', action='store_true')
 
     nii.add_argument("input", help=input_str, type=str)
     nii.add_argument("-b", "--bids", help=bids_opt, action='store_true')
     nii.add_argument("-o", "--output", help=output_fnm_str, type=str, default=False)
-    nii.add_argument("-r", "--recoid", help="RECO ID (default=1), option to specify a particular reco to convert",
+    nii.add_argument("-r", "--recoid", help="RECO ID (default=1), "
+                                            "option to specify a particular reconstruction id to convert",
                      type=int, default=1)
     nii.add_argument("-s", "--scanid", help="Scan ID, option to specify a particular scan to convert.", type=str)
-    nii.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
+    nii.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
+    nii.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
+    nii.add_argument("--ignore-rescale", help='remove slope and offset value from header', action='store_true')
 
     niiall.add_argument("input", help=input_dir_str, type=str)
     niiall.add_argument("-o", "--output", help=output_dir_str, type=str)
     niiall.add_argument("-b", "--bids", help=bids_opt, action='store_true')
-    niiall.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
+    niiall.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
+    niiall.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
+    niiall.add_argument("--ignore-rescale", help='remove slope and offset value from header', action='store_true')
 
     bids_helper.add_argument("input", help=input_dir_str, type=str)
     bids_helper.add_argument("output", help="output BIDS datasheet filename (.xlsx)", type=str)
@@ -73,7 +80,9 @@ def main():
     bids_convert.add_argument("datasheet", help="input BIDS datahseet filename", type=str)
     bids_convert.add_argument("-j", "--json", help="input JSON syntax template filename", type=str, default=False)
     bids_convert.add_argument("-o", "--output", help=output_dir_str, type=str, default=False)
-    bids_convert.add_argument("--ignore_slope", help='remove slope value from header', action='store_true')
+    bids_convert.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
+    bids_convert.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
+    bids_convert.add_argument("--ignore-rescale", help='remove slope and offset value from header', action='store_true')
 
     args = parser.parse_args()
 
@@ -97,8 +106,19 @@ def main():
         root = MainWindow()
         if ipath != None:
             root._path = ipath
-            if args.ignore_slope is True:
+            if not args.ignore_rescale:
+                if args.ignore_slope:
+                    root._ignore_slope = True
+                else:
+                    root._ignore_slope = False
+                if args.ignore_offset:
+                    root._ignore_offset = True
+                else:
+                    root._ignore_offset = False
+            else:
                 root._ignore_slope = True
+                root._ignore_offset = True
+
[…]
