### 2020-06-04 924a9a7
[resolve issue] #3 and #7
