### 2020-06-04 686036a
[resolve issue] #9 remove position correction (Bruker performed internally for Visu parameters)
