### 2020-06-04 1d756a7
[patch] #8 DTI will have default motadality as 'dwi' on BIDS helper sheet. Correct bvec to have 3 rows instead of 3 columns
