+++ README.md
@@ -58,21 +58,21 @@ to help benchmark testing of Bruker converter.
 GNU General Public License v3.0
 
 ### Citing BrkRaw
-Lee, Sung-Ho, Ban, Woomi, & Shih, Yen-Yu Ian. (2020, May 19). BrkRaw/bruker: BrkRaw v0.3.3rc0 (Version 0.3.3rc0). 
-Zenodo. http://doi.org/10.5281/zenodo.3833397
+Lee, Sung-Ho, Ban, Woomi, & Shih, Yen-Yu Ian. (2020, June 4). BrkRaw/bruker: BrkRaw v0.3.3 (Version 0.3.3). 
+Zenodo. http://doi.org/10.5281/zenodo.3877179
 
 **BibTeX**
 ```
-@software{lee_sung_ho_2020_3833397,
+@software{lee_sung_ho_2020_3877179,
   author       = {Lee, Sung-Ho and
                   Ban, Woomi and
                   Shih, Yen-Yu Ian},
-  title        = {BrkRaw/bruker: BrkRaw v0.3.3rc0},
-  month        = may,
+  title        = {BrkRaw/bruker: BrkRaw v0.3.3},
+  month        = june,
   year         = 2020,
   publisher    = {Zenodo},
-  version      = {0.3.3rc0},
-  doi          = {10.5281/zenodo.3833397},
-  url          = {https://doi.org/10.5281/zenodo.3833397}
+  version      = {0.3.3},
+  doi          = {10.5281/zenodo.3877179},
+  url          = {https://doi.org/10.5281/zenodo.3877179}
 }
 ```
