+++ brkraw/scripts/brkraw.py
@@ -170,6 +170,9 @@ def main():
                     output_fname = '{}-{}-{}'.format(output, str(scan_id).zfill(2), reco_id)
                     try:
                         study.save_as(scan_id, reco_id, output_fname, slope=slope, offset=offset)
+                        method = study._pvobj._method[scan_id].parameters['Method']
+                        if re.search('dti', method, re.IGNORECASE):
+                            study.save_bdata(scan_id, output_fname)
                         if args.bids:
                             study.save_json(scan_id, reco_id, output_fname)
                         print('NifTi file is genetared... [{}]'.format(output_fname))
