+++ brkraw/lib/loader.py
@@ -998,7 +998,11 @@ class BrukerLoader():
         orient_info = self._get_orient_info(visu_pars, method)
         slice_orient_map = {0: 'sagital', 1: 'coronal', 2: 'axial'}
         num_slice_packs = slice_info['num_slice_packs']
-        subj_pose = orient_info['subject_position']
+        version = get_value(visu_pars, 'VisuVersion')
+        if version == 1:  # PV 5.1 required the position correction
+            subj_pose = orient_info['subject_position']
+        else:             # PV 6.01 does not
+            subj_pose = None
         subj_type = orient_info['subject_type']
 
         if num_slice_packs > 1:
