### 2020-06-22 #2277 @PeerHerholz
Hi @richardjgowers,

I just did the first round of my review. I want to provide precise feedback to the authors regarding the points I did not check and thus need to be addressed in my opinion. This should be done here, correct?
