### 2020-06-22 #2277 @richardjgowers
Sure or raise them as issues on the repo if that’s more appropriate.

On Mon, Jun 22, 2020 at 16:30, Peer Herholz <notifications@github.com>
wrote:

> Hi @richardjgowers <https://github.com/richardjgowers>,
>
> I just did the first round of my review. I want to provide precise
> feedback to the authors regarding the points I did not check and thus need
> to be addressed in my opinion. This should be done here, correct?
>
