+++ brkraw/__init__.py
@@ -1,6 +1,6 @@
 from .lib import *
 
-__version__ = '0.3.4b0'
+__version__ = '0.3.4'
 __all__ = ['BrukerLoader', '__version__']
 
 
