+++ README.md
@@ -1,7 +1,7 @@
 [![DOI](https://zenodo.org/badge/245546149.svg)](https://zenodo.org/badge/latestdoi/245546149)
 
 ## BrkRaw: A comprehensive tool to access raw Bruker Biospin MRI data
-#### Version: 0.3.3
+#### Version: 0.3.4
 
 ### Description
 
@@ -50,10 +50,10 @@ please visit our [GitPage](https://brkraw.github.io).
 - Drs. Chris Rorden and Sebastiano Ferraris: The pioneers related this project who had been inspired the developer
  through their great tools including [dcm2niix](https://github.com/rordenlab/dcm2niix) and 
  [bruker2nifti](https://github.com/SebastianoF/bruker2nifti), as well as their comments to improve this project. 
-- Drs. Mikael Naveau and Gabriel A. Devenyi: The publisher of 
+- Dr. Mikael Naveau: The publisher of 
 [bruker2nifti_qa](https://gitlab.com/naveau/bruker2nifti_qa), the set of data 
 to help benchmark testing of Bruker converter.
-- 
+- Dr. Gabriel A. Devenyi: The vast contributions to refinement of module functionality and troubleshooting.
 
 ### License:
 GNU General Public License v3.0
@@ -65,16 +65,16 @@ Zenodo. http://doi.org/10.5281/zenodo.3877179
 
 **BibTeX**
 ```
-@software{lee_sung_ho_2020_3877179,
+@software{lee_sung_ho_2020_3907018,
   author       = {Lee, Sung-Ho and
                   Ban, Woomi and
                   Shih, Yen-Yu Ian},
-  title        = {BrkRaw/bruker: BrkRaw v0.3.3},
-  month        = june,
+  title        = {BrkRaw/bruker: BrkRaw v0.3.4},
+  month        = jun,
   year         = 2020,
   publisher    = {Zenodo},
-  version      = {0.3.3},
-  doi          = {10.5281/zenodo.3877179},
-  url          = {https://doi.org/10.5281/zenodo.3877179}
+  version      = {0.3.4},
+  doi          = {10.5281/zenodo.3907018},
+  url          = {https://doi.org/10.5281/zenodo.3907018}
 }
 ```
