### 2020-06-26 BrkRaw/brkraw#13 @PeerHerholz
review comments based on JOSS submissin
Hello everyone,

my name is Peer and I'm one of the reviewers of the manuscript you've submitted to [JOSS](https://github.com/openjournals/joss-reviews/issues/2277) regarding this repository and tool.

Based on my first round of review, I would like to provide some feedback regarding the review points I did not checked and I think need just a little more work to bring this already great manuscript home. The respective points will be outlined below. If you have the impression that my feedback is not constructive, my wording to harsh, I misunderstood something and/or have any other questions, please let me know.

### Documentation

- [x] the `statement of need` could be more precisely outlined and described on the landing page following JOSS guidelines: "...what problems the software is designed to solve and who the target audience is?"
- [x] while there's already information regarding different use cases and functionality, the respective parts could/should be demonstrated on data: a step-step walkthrough would be great, converting real world example data
- [x] regarding the `API`: I checked your respective [jupyter notebook](https://github.com/brkraw/bruker/blob/master/examples/BrkRaw_PythonAPI.ipynb) and would like to suggest to add more information and description to this notebook, explaining things a bit more in depth. Furthermore, you could think about adding this notebook in an interactive version (through the amazing [binder project](https://mybinder.org/)) via a badge to your documentation.

### Software
- [x] While I know that your software is not "classic" when it comes to `automated tests`, I think adding some would be great in order to ensure functionality in the future. Maybe certain parts of the `API` could be tested, that is if data is converted as expected?

### Repository and Community
- [x] It would be great if you could include resources that make it easier for the community to interact with you as developers. This could/should entail an `Issue template`, `Contributing guidelines` and information regarding where to get `support`.

### General
- [x] The writing of the paper could be improved (a few missing words/articles here and there, present/past, etc.).
- [x] The `Naveau:2019` reference is not correctly rendered. Could you please check if that's maybe a typo in your submission or a problem on `JOSS`' end?

Again, if you have any questions regarding my comments, please don't hesitate to let me know. If you consider addressing them, I think it would be great to submit one PR for each comment to ensure traceability during the review process and beyond. I would like to thank you again for submitting this interesting manuscript.

I hope you and your loved ones are doing ok in these difficult times.

Best regards, Peer
