### 2020-06-26 #2277 @PeerHerholz
A opened an issue outlining my comments [here](https://github.com/BrkRaw/bruker/issues/13).
