### 2020-06-28 BrkRaw/bruker#13 @dvm-shlee
@PeerHerholz Thank you for the useful and detailed comments. We will update the package according to each item and will generate PR for each item as well. Regarding building community, as up now.
