### 2020-06-28 BrkRaw/bruker#13 @PeerHerholz
Hi @dvm-shlee,

cool, thanks a lot. I very much look forward to it!
