### 2020-07-05 BrkRaw/brkraw#14 @mih
Unsubstantiated claim re market share OR citation needed
Hey, this issue is part of my review of your JOSS submission at https://github.com/openjournals/joss-reviews/issues/2277

A bit of a nit pick, but in the paper you state

> Bruker Biospin MRI (the leading preclinical MRI system)

I understand that being the leader indirectly justifies the need for BrkRaw. However, "leading" has many connotations, and it would certainly be sufficient to claim that quite a few Bruker systems are in use in the field, without declaring them a "leader" of some kind (without a citation that qualifies this further).
