+++ paper/paper.md
@@ -47,7 +47,7 @@ for converting raw data directly into a NIfTI or ANALYZE format
 
 Although the converters significantly improve analysis workflow by direct converting into software friendly file format,
 the software introduced so far has not provided the feature to access the data immediately without conversion.
-Here we present the 'BrkRaw' python module, a more comprehensive preclinical tool for accessing and utilizing
+Here we present the 'BrkRaw' Python module, a more comprehensive preclinical tool for accessing and utilizing
 raw Bruker Biospin MRI data. The module has been built up upon robust low-level Python Application Programming Interface 
 (API), allowing direct raw data access without conversion to provide the advanced and easy-to-use features 
 for data analysis. 
