### 2020-07-05 BrkRaw/brkraw#16 @mih
Incompatible with Python 3.8?
Hey, this issue is part of my review of your JOSS submission at https://github.com/openjournals/joss-reviews/issues/2277

I am running Debian bullseye with Python 3.8.3. I cannot install brkraw using this version, because 3.8 is not a supported version.

```
      python_requires='>3.5, <3.8',
```

Going through the source code I cannot easily find a reason for that. Is 3.8 incompatible? Do you have plans to support 3.8?
