### 2020-07-05 BrkRaw/bruker#17 @mih
Could SimpleITK become an optional dependency?
Hey, this issue is part of my review of your JOSS submission at https://github.com/openjournals/joss-reviews/issues/2277

SimpleITK is a quite heavy (>40MB) and needs to be compiled (in contrast to the pure Python NiBabel). Moreover, it is only used in a single location `lib/loader.py:get_sitkimg()` via an internal import.

Could this be made an optional dependency and safe non-ITK users the potential complications?
