### 2020-07-05 BrkRaw/bruker#18 @mih
LICENSE file contains more than one license (merge-conflict left-over)
Hey, this issue is part of my review of your JOSS submission at openjournals/joss-reviews#2277

It starts like this:

```
<<<<<<< HEAD
                    GNU GENERAL PUBLIC LICENSE
=======
                   GNU LESSER GENERAL PUBLIC LICENSE
>>>>>>> 293c218a6d432b192fba6d65b22d9dc843cb1c7e
```

but the entire file seems to be affected.
