### 2020-07-05 4272971
Fix typo
Spotted during my review of your JOSS submission at https://github.com/openjournals/joss-reviews/issues/2277
