### 2020-07-05 BrkRaw/brkraw#20 @mih
Contribute low-level code to NiBabel?
Hey, this issue is part of my review of your JOSS submission at openjournals/joss-reviews#2277

This package provides valuable functionality that is not available in NiBabel. However, NiBabel, arguably, is the key package for neuroimaging file format support in the Python ecosystem. It has more than 80 contributors so far, and a solid track record for keeping essential data IO procedures alive and well for a long period of time. Given that much academic software shares the lifecycle of PhD students, I consider it important to integrate as much code as possible in community-maintained projects to improve the overall resilience of the research software ecosystem.

BrkRaw's scope is clearly wider than what would be a suitable contribution for NiBabel. However, NiBabel is already a dependency, so moving low-level code into this core library would not complicate deployment for BrkRaw users.

Did you consider such a contribution?
