### 2020-07-05 BrkRaw/brkraw#21 @mih
How can correct function be verified?
Hey, this issue is part of my review of your JOSS submission at openjournals/joss-reviews#2277

The code repository does not seem to contain unittests or another kind of tests, and I did not find instructions on how to verify correct behavior for a particular deployment. There is also no CI set up on GitHub. Given that correct behavior of research data IO implementations is crucial (considering their position at the starting point of any analysis), it would be instrumental to be able to verify correct behavior.

This issue is somewhat related to #20 (integration with NiBabel, which comes with an extensive test suite, and test helpers and strategies that could be useful in the context of BrkRaw).
