### 2020-07-05 BrkRaw/bruker#17 @gdevenyi
https://github.com/BrkRaw/bruker/blob/master/brkraw/lib/loader.py#L356-L405
