### 2020-07-05 BrkRaw/bruker#21 @gdevenyi
I suggest perhaps that we could validate against a NIFTI export from Paravision360, however I don't have it and I don't know anyone who does who could provide some example data.

Having said that, I'm not sure I trust Bruker to actually properly implement a NIFTI conversion anyways.
