### 2020-07-05 BrkRaw/bruker#20 @gdevenyi
Does nibabel provide DICOM-to-NIFTI export? That would be equivalent to adding this code into nibabel, otherwise, I'm not sure I understand the suggestion.
