### 2020-07-06 BrkRaw/brkraw#22 @mwegrzyn
Review Comments JOSS, Reviewer Martin Wegrzyn
Dear @dvm-shlee et al.,

my name is Martin Wegrzyn and I am one of the reviewers assigned to your submission to JOSS. Thank you for your work on BrkRaw. The following are some comments I have regarding my first impression of the package and the manuscript:

**code & examples**

- [ ]  I tried to run the notebook ```examples/BrkRaw_PythonAPI.ipynb``` but got a number of error messages and some inconsistencies with the outputs of the notebook you uploaded.

* Cell 8: Error: ```“AttributeError: module 'brkraw.lib.reference' has no attribute 'METADATA_FILED_INFO'”```
* Cells 15, 20 ```“AttributeError: 'NoneType' object has no attribute 'T'”```
* Cell 37: ``` “TypeError: get_dataobj() got an unexpected keyword argument 'slp_correct'”```
* Cell 39 (and others) different output than the one on GitHub: e.g. your affine matrix is

[[  0.4    -0.      0.004 -14.381]
 [ -0.     -0.4     0.     14.968]
 [ -0.004   0.      0.4     1.824]
 [  0.      0.      0.      1.   ]]
```
mine, when running the code, is

[[-0.399978	-0.000000	-0.004234	14.381493]
 [0.000000	-0.400000	-0.000000	14.968180]
 [0.004234	0.000000	-0.399978	-1.824436]
 [0.000000	0.000000	0.000000	1.000000]]
```

Hence, there are some changes in the signs (last column). Could this be a problem e.g. with left-right flips of the data?

compare: https://gist.github.com/mwegrzyn/1c8113631a4ab3fe19cf1838f3356b56

- [ ] regarding the example data, I wanted to ask whether there would be a possibility for me to confirm where left and right is in the image?
- [ ] also regarding the example data, the BOLD/EPI images seem to be 3D-files (single volume), not 4D timeseries; is this correct or should I be able to see different volumes? I think it would be useful to have 4D files for testing purposes. Thank you!

- [ ] the BIDS-conversion worked very well, however, it only worked for me if I did not use the .json file and only used the .xlsx file. The command
brkraw bids_convert 20190724_114946_BRKRAW_1_1 dataset_description.xlsx -r dataset_description.json
```
did not work:
brkraw: error: unrecognized arguments: -r dataset_description.json”
```
only using
brkraw bids_convert 20190724_114946_BRKRAW_1_1 dataset_description.xlsx”
```
did work fine!

**Manuscript**

- [ ] regarding the motivation of the project, when reading the manuscript I get the feeling that you argue that nifiti-conversion is useful; also, your package has the ability to convert to nifti; on the other hand, the companion website mentions that nifiti conversion is a "legacy" function. Hence, I am unsure what the nature of "BrkRaw" is and how much the “raw” handling of the data is desirable from your point of view. E.g. _“The module has been built up upon robust low-level Python Application Programming Interface (API), allowing direct raw data access without conversion to provide the advanced and easy-to-use features for data analysis.”_; here, I would like to learn more about the analysis on raw data mentioned and why Nifiti conversion is nevertheless such a central part of the package. Thank you.

- [ ] I think it is fair to say that the manuscript would need proofreading. Also, there are some sentences where the meaning is a bit unclear. For example at the end of 1st paragraph: _“However, the preclinical MRI research using laboratory animals or objects does not require the DICOM conversion step and is rather inefficient.”_ > this sounds like preclinical MRI is inefficient, while I assume that the intended meaning is that the DICOM conversion step would be inefficient for preclinical MRI?

- [ ] Figures: Overall, I found the figures difficult to understand. Maybe more extensive figure legends or in-text descriptions could be helpful (although I saw in other JOSS papers, that figure descriptions are often as brief as yours. Still, it might be worth considering)
- [ ] Figure 1: Could you explain what we see in the _“between dataset comparison”_ and how we can visually assess successful conversion?
- [ ] Figure 4: the contents of the four subplots are not readable. I would be great if you could increase the resolution of the figure. Thank you.

I hope you find these comments to be helpful. Please let me know if you have any questions regarding the points I raised. Thank you for your time and consideration.

Martin Wegrzyn
