### 2020-07-06 #2277 @mwegrzyn
Dear colleagues,

my review of BrkRaw can be found here: https://github.com/BrkRaw/bruker/issues/22

Best regards,
Martin Wegrzyn
