### 2020-07-06 BrkRaw/bruker#16 @dvm-shlee
@mih Thank you for your comments,

The supported versions are listed on the documentation page: https://brkraw.github.io/docs/gs_inst.html as 3.6 and 3.7 only
because those are only the versions I've fully tested so far.
There are few issues supporting 3.8 at this point due to some syntax changes, which I experience some unexpected bugs.
So, I just blocked it intentionally to prevent this issue.

I would like to support 3.8 later but prefer to do it immediately, but during revising the paper and source code, I will integrate the test scripts which commented by (@PeerHerholz) to support recent versions of Python and clarify the supported python version python on the README.md.
