### 2020-07-06 BrkRaw/bruker#20 @dvm-shlee
@mih Thank you for your comment, as @gdevenyi pointed out, the low-level code aims to load Bruker raw data without conversion. The feature I added to support NiBabel NifTi object and Simple ITK image object helps the user can access the data in their convenient way without conversion.

Since the NiBabel is not designed for the raw data conversion, I don't think the low-level code can be integrated into the NiBabel. I hope this can cover your question!
