### 2020-07-06 BrkRaw/bruker#17 @dvm-shlee
@mih This is the good point, I will try to find a way to make it optional and revise the script.
