### 2020-07-06 BrkRaw/bruker#21 @dvm-shlee
@mih Thank you for your valuable comment! As @gdevenyi pointed, most researchers do not believe the fidelity of the converter even Bruker provided themselves. So it's somewhat hard to guarantee that this library can behave well for all data. For this reason, I was trying to interact directly with the user to deal with the issue, so improve the fidelity of the converter. However, I strongly agree that the testing script is needed to make it more productive for maintaining this module.

I'm considering adding test-scripts to check if the BrkRaw module can convert the meta-data, which has been tested, correctly. This will not guarantee the fidelity of the module, but at least I'm expecting this can help to manage the version well for future collaborative work with other researchers.
