### 2020-07-14 BrkRaw/bruker#21 @gdevenyi
Our group is preparing a large dataset of PV5/PV6 scans of a phantom with all the default sequences, tweaking a large variety of parameters as well as subject type and orientation. We'll be happy to contribute this as a validation set for this.
