### 2020-07-14 BrkRaw/bruker#21 @dvm-shlee
@gdevenyi  Thank you for your suggestion, I appreciate your heavy contribution to this project! I just sent an invitation to this project in case you interested in contributing as an author as well! Thank you so much!
