### 2020-07-24 5dfd5b3
[patch] Issue #12, PV5.1 JSON parser, related to multiple phase encoding direction.
