+++ brkraw/lib/utils.py
@@ -1,125 +1,12 @@
+from shleeh.errors import UnexpectedError
+
 from .reference import *
 import re
 import os
-import math
 import numpy as np
 from collections import OrderedDict
 from functools import partial, reduce
 from copy import copy as cp
-from nibabel.affines import to_matvec, from_matvec
-
-
-def apply_affine(matrix, affine):
-    return affine.dot(matrix)
-
-
-def apply_rotate(matrix, rad_x=0, rad_y=0, rad_z=0):
-    ''' axis = x or y or z '''
-    rmat = dict(x = np.array([[1, 0, 0],
-                              [0, np.cos(rad_x), -np.sin(rad_x)],
-                              [0, np.sin(rad_x), np.cos(rad_x)]]).astype('float'),
-                y = np.array([[np.cos(rad_y), 0, np.sin(rad_y)],
-                              [0, 1, 0],
-                              [-np.sin(rad_y), 0, np.cos(rad_y)]]).astype('float'),
-                z = np.array([[np.cos(rad_z), -np.sin(rad_z), 0],
-                              [np.sin(rad_z), np.cos(rad_z), 0],
-                              [0, 0, 1]]).astype('float'))
-    af_mat, af_vec = to_matvec(matrix)
-    rotated_mat = rmat['z'].dot(rmat['y'].dot(rmat['x'].dot(af_mat)))
-    rotated_vec = rmat['z'].dot(rmat['y'].dot(rmat['x'].dot(af_vec)))
-    return from_matvec(rotated_mat, rotated_vec)
-
-
-def reversed_pose_correction(pose, rmat, distance):
-    reversed_pose = rmat.dot(pose)
-    reversed_pose[-1] += distance
-    corrected_pose = rmat.T.dot(reversed_pose)
-    return corrected_pose
-
-
-def build_affine_from_orient_info(resol, rmat, pose,
-                                  subj_pose, subj_type, slice_orient):
-    if slice_orient in ['axial', 'sagital']:
-        resol = np.diag(np.array(resol))
-    else:
-        resol = np.diag(np.array(resol) * np.array([1, 1, -1]))
-    rmat = rmat.T.dot(resol)
-
-    affine = from_matvec(rmat, pose)
-
-    # convert space from image to subject
-    # below positions are all reflext human-based position
-    if subj_pose:
-        if subj_pose == 'Head_Supine':
-            affine = apply_rotate(affine, rad_z=np.pi)
-        elif subj_pose == 'Head_Prone':
-            pass
-        # From here, not urgent, extra work to determine correction matrix needed.
-        elif subj_pose == 'Head_Left':
-            affine = apply_rotate(affine, rad_z=np.pi/2)
-        elif subj_pose == 'Head_Right':
-            affine = apply_rotate(affine, rad_z=-np.pi/2)
-        elif subj_pose in ['Foot_Supine', 'Tail_Supine']:
-            affine = apply_rotate(affine, rad_x=np.pi)
-        elif subj_pose in ['Foot_Prone', 'Tail_Prone']:
-            affine = apply_rotate(affine, rad_y=np.pi)
-        elif subj_pose in ['Foot_Left', 'Tail_Left']:
-            affine = apply_rotate(affine, rad_z=np.pi/2)
-        elif subj_pose in ['Foot_Right', 'Tail_Right']:
-            affine = apply_rotate(affine, rad_z=-np.pi/2)
-        else:  # in case Bruker put additional value for this header
-            raise Exception(ERROR_MESSAGES['NotIntegrated'])
-
-    if subj_type != 'Biped':
-        # correct subject space if not biped (human or non-human primates)
[…]
