+++ brkraw/lib/orient.py
@@ -1,0 +1,191 @@
+import math
+from copy import copy as cp
+
+import numpy as np
+from nibabel.affines import from_matvec, to_matvec
+
+from bruker.brkraw.lib.reference import ERROR_MESSAGES
+
+
+def build_affine_from_orient_info(resol, rmat, pose,
+                                  subj_pose, subj_type, slice_orient):
+    if slice_orient in ['axial', 'sagital']:
+        resol = np.diag(np.array(resol))
+    else:
+        resol = np.diag(np.array(resol) * np.array([1, 1, -1]))
+    rmat = rmat.T.dot(resol)
+
+    affine = from_matvec(rmat, pose)
+
+    # convert space from image to subject
+    # below positions are all reflect human-based position
+    if subj_pose:
+        if subj_pose == 'Head_Supine':
+            affine = apply_rotate(affine, rad_z=np.pi)
+        elif subj_pose == 'Head_Prone':
+            pass
+        # From here, not urgent, extra work to determine correction matrix needed.
+        elif subj_pose == 'Head_Left':
+            affine = apply_rotate(affine, rad_z=np.pi/2)
+        elif subj_pose == 'Head_Right':
+            affine = apply_rotate(affine, rad_z=-np.pi/2)
+        elif subj_pose in ['Foot_Supine', 'Tail_Supine']:
+            affine = apply_rotate(affine, rad_x=np.pi)
+        elif subj_pose in ['Foot_Prone', 'Tail_Prone']:
+            affine = apply_rotate(affine, rad_y=np.pi)
+        elif subj_pose in ['Foot_Left', 'Tail_Left']:
+            affine = apply_rotate(affine, rad_z=np.pi/2)
+        elif subj_pose in ['Foot_Right', 'Tail_Right']:
+            affine = apply_rotate(affine, rad_z=-np.pi/2)
+        else:  # in case Bruker put additional value for this header
+            raise Exception(ERROR_MESSAGES['NotIntegrated'])
+
+    if subj_type != 'Biped':
+        # correct subject space if not biped (human or non-human primates)
+        # not sure this rotation is independent with subject pose, so put here instead last
+        affine = apply_rotate(affine, rad_x=-np.pi/2, rad_y=np.pi)
+    return affine
+
+
+def reversed_pose_correction(pose, rmat, distance):
+    reversed_pose = rmat.dot(pose)
+    reversed_pose[-1] += distance
+    corrected_pose = rmat.T.dot(reversed_pose)
+    return corrected_pose
+
+
+def is_rotation_matrix(matrix):
+    t_matrix = np.transpose(matrix)
+    should_be_identity = np.dot(t_matrix, matrix)
+    i = np.identity(3, dtype=matrix.dtype)
+    n = np.linalg.norm(i - should_be_identity)
+    return n < 1e-6
+
+
+def apply_flip(matrix, axis, mat=True, vec=True):
+    '''axis = x or y or z'''
+    flip_idx = dict(x=0, y=1, z=2)
+    orig_mat, orig_vec = to_matvec(matrix)
+
+    aff_mat = np.ones(3)
+    aff_mat[flip_idx[axis]] = -1
+    aff_mat = np.diag(aff_mat)
+    if mat:
+        flip_mat = aff_mat.dot(orig_mat)
+    else:
+        flip_mat = orig_mat
+    if vec:
+        flip_vec = aff_mat.dot(orig_vec)
+    else:
+        flip_vec = orig_vec
+    return from_matvec(flip_mat, flip_vec)
+
+
+def calc_eulerangle(matrix):
+    assert (is_rotation_matrix(matrix))
+
+    sy = math.sqrt(matrix[0, 0] * matrix[0, 0] + matrix[1, 0] * matrix[1, 0])
+    singular = sy < 1e-6
+    if not singular:
[…]

[…]
