+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,1 +1,8 @@
-# temporary file
+Fixes #.
+
+Changes proposed in this pull request:
+-
+-
+-
+
+@BrkRaw/bruker
