### 2020-07-31 0489ff2
Delete pull_request_template.md
