+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,8 +1,0 @@
-Fixes #.
-
-Changes proposed in this pull request:
--
--
--
-
-@BrkRaw/bruker
