+++ .github/pull_request_template.md
@@ -1,0 +1,8 @@
+Fixes #.
+
+Changes proposed in this pull request:
+-
+-
+-
+
+@BrkRaw/bruker
