### 2020-07-31 dfdf207
[patch] #27 brkraw.py 'tonii_all' improve handling invalid input.
