+++ brkraw/scripts/brkraw.py
@@ -2,8 +2,8 @@
 from shleeh import *
 from shleeh.errors import *
 
-from bruker.brkraw.lib.utils import mkdir
 from .. import BrukerLoader, __version__
+from ..lib.utils import set_rescale, save_meta_files, mkdir
 import argparse
 import os, re
 
@@ -75,7 +75,8 @@ def main():
     bids_convert.add_argument("-o", "--output", help=output_dir_str, type=str, default=False)
     bids_convert.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
     bids_convert.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
-    bids_convert.add_argument("--ignore-rescale", help='remove slope and offset values from header', action='store_true')
+    bids_convert.add_argument("--ignore-rescale", help='remove slope and offset values from header',
+                              action='store_true')
 
     args = parser.parse_args()
 
@@ -125,19 +126,7 @@ def main():
         scan_id = args.scanid
         reco_id = args.recoid
         study = BrukerLoader(path)
-
-        if not args.ignore_rescale:
-            if args.ignore_slope:
-                slope = None
-            else:
-                slope = False
-            if args.ignore_offset:
-                offset = None
-            else:
-                offset = False
-        else:
-            slope = None
-            offset = None
+        slope, offset = set_rescale(args)
 
         if args.output:
             output = args.output
@@ -149,99 +138,82 @@ def main():
                 scan_id = int(scan_id)
                 reco_id = int(reco_id)
                 study.save_as(scan_id, reco_id, output_fname, slope=slope, offset=offset)
-                method = study._pvobj._method[scan_id].parameters['Method']
-                if re.search('dti', method, re.IGNORECASE):
-                    study.save_bdata(scan_id, output_fname)
-                if args.bids:
-                    study.save_json(scan_id, reco_id, output_fname)
+                save_meta_files(study, args, scan_id, reco_id, output_fname)
                 print('NifTi file is generated... [{}]'.format(output_fname))
-            except Exception:
-                print('Failed..')
+            except:
+                print(f'Conversion failed: ScanID:{str(scan_id)}, RecoID:{str(reco_id)}')
         else:
             for scan_id, recos in study._pvobj.avail_reco_id.items():
                 for reco_id in recos:
                     output_fname = '{}-{}-{}'.format(output, str(scan_id).zfill(2), reco_id)
                     try:
                         study.save_as(scan_id, reco_id, output_fname, slope=slope, offset=offset)
-                        method = study._pvobj._method[scan_id].parameters['Method']
-                        if re.search('dti', method, re.IGNORECASE):
-                            study.save_bdata(scan_id, output_fname)
-                        if args.bids:
-                            study.save_json(scan_id, reco_id, output_fname)
+                        save_meta_files(study, args, scan_id, reco_id, output_fname)
                         print('NifTi file is genetared... [{}]'.format(output_fname))
-                    except Exception as e:
-                        print('Failed..'.format(e))
+                    except:
+                        print(f'Conversion failed: ScanID:{str(scan_id)}, RecoID:{str(reco_id)}')
 
     elif args.function == 'tonii_all':
-        path = args.input
-
-        if not args.ignore_rescale:
-            if args.ignore_slope:
-                slope = None
-            else:
-                slope = False
-            if args.ignore_offset:
-                offset = None
-            else:
-                offset = False
-        else:
-            slope = None
-            offset = None
-
         from os.path import join as opj, isdir, isfile
+
+        path = args.input
+        slope, offset = set_rescale(args)
+        invalid_error_message = f'[Error] Invalid input path: {path}\n'
+        empty_folder = f'        The input path does not contain any raw data.'
+        wrong_target = f'        The input path indicates raw data itself. \n' \
[…]
