### 2020-07-31 2019fa8
[patch] #17 setup.py 'SimpleITK' now became optional dependency, documentation updated on https://brkraw.github.io/docs/gs_inst.html
