### 2020-07-31 86d3e3c
[patch] #17 loader.py Raise error message when 'get_sitkimg' method used without optional dependency 'SimpleITK'
