### 2020-07-31 cce87ce
#18 Correct LICENSE file
