+++ README.md
@@ -59,6 +59,10 @@ to help benchmark testing of Bruker converter.
 ### License:
 GNU General Public License v3.0
 
+### How to get Support
+If you are experiencing any problem or have questions, please report it through 
+[Issues](https://github.com/BrkRaw/bruker/issues)
+
 ### Citing BrkRaw
 Lee, Sung-Ho, Ban, Woomi, & Shih, Yen-Yu Ian. (2020, June 4). BrkRaw/bruker: BrkRaw v0.3.3 (Version 0.3.3). 
 Zenodo. http://doi.org/10.5281/zenodo.3877179
