### 2020-07-31 84fe835
[update] paper.md #13 Rendering issue of (Naveau:2019)
