### 2020-07-31 b228518
[organizing] #13 Reorganization of Tutorial and Example dataset for setting up interactive Jupyter notebook and automate test
