+++ .github/workflows/ci_workflow.yml
@@ -20,7 +20,7 @@ jobs:
     - name: Set up python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
-        python-version: $${{ matrix.python-version }}
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
