+++ brkraw/lib/loader.py
@@ -165,8 +165,7 @@ class BrukerLoader():
             f = fg_info['frame_size']
             if isinstance(data_slp, list):
                 if f != len(data_slp):
-                    raise UnexpectedError(message='data_slp mismatch;'
-                                                  f'{ISSUE_REPORT}')
+                    raise UnexpectedError(message='data_slp mismatch;{}'.format(ISSUE_REPORT))
                 else:
                     if dim == 2:
                         x, y = matrix_size[:2]
@@ -175,8 +174,7 @@ class BrukerLoader():
                         x, y, z = matrix_size[:3]
                         _dataobj = dataobj.reshape([f, x * y * z]).T
                     else:
-                        raise UnexpectedError(message='Unexpected frame shape on DTI image;'
-                                                      f'{ISSUE_REPORT}')
+                        raise UnexpectedError(message='Unexpected frame shape on DTI image;{}'.format(ISSUE_REPORT))
                 dataobj = (_dataobj * data_slp).T
             else:
                 dataobj = dataobj * data_slp
@@ -186,8 +184,7 @@ class BrukerLoader():
             f = fg_info['frame_size']
             if isinstance(data_off, list):
                 if f != len(data_off):
-                    raise UnexpectedError(message='data_off mismatch;'
-                                                  f'{ISSUE_REPORT}')
+                    raise UnexpectedError(message='data_off mismatch;{}'.format(ISSUE_REPORT))
                 else:
                     if dim == 2:
                         x, y = matrix_size[:2]
@@ -196,8 +193,7 @@ class BrukerLoader():
                         x, y, z = matrix_size[:3]
                         _dataobj = dataobj.reshape([f, x * y * z]).T
                     else:
-                        raise UnexpectedError(message='Unexpected frame shape on DTI image;'
-                                                      f'{ISSUE_REPORT}')
+                        raise UnexpectedError(message='Unexpected frame shape on DTI image;{}'.format(ISSUE_REPORT))
                 dataobj = (_dataobj + data_off).T
             else:
                 dataobj = dataobj + data_off
@@ -231,8 +227,7 @@ class BrukerLoader():
             elif group_id[0] in ['FG_DIFFUSION', 'FG_DTI', 'FG_MOVIE', 'FG_COIL', 'FG_CYCLE', 'FG_COMPLEX']:
                 dataobj = swap_slice_axis(group_id, dataobj)
             else:
-                raise UnexpectedError(message='Unexpected frame group combination;'
-                                              f'{ISSUE_REPORT}')
+                raise UnexpectedError(message='Unexpected frame group combination;{}'.format(ISSUE_REPORT))
         return dataobj
 
     def get_fid(self, scan_id):
@@ -426,15 +421,15 @@ class BrukerLoader():
 
     def _inspect_ids(self, scan_id, reco_id):
         if scan_id not in self._avail.keys():
-            print(f'[Error] Invalid Scan ID.\n'
-                  f'  - Your input: {scan_id}\n'
-                  f'  - Available Scan IDs: {list(self._avail.keys())}')
+            print('[Error] Invalid Scan ID.\n'
+                  '  - Your input: {}\n'
+                  '  - Available Scan IDs: {}'.format(scan_id, list(self._avail.keys())))
             raise ValueError
         else:
             if reco_id not in self._avail[scan_id]:
-                print(f'[Error] Invalid Reco ID.\n'
-                      f'  - Your input: {reco_id}\n'
-                      f'  - Available Reco IDs: {self._avail[scan_id]}')
+                print('[Error] Invalid Reco ID.\n'
+                      '  - Your input: {}\n'
+                      '  - Available Reco IDs: {}'.format(reco_id, self._avail[scan_id]))
                 raise ValueError
 
[…]
