+++ brkraw/scripts/brkraw.py
@@ -141,7 +141,7 @@ def main():
                 save_meta_files(study, args, scan_id, reco_id, output_fname)
                 print('NifTi file is generated... [{}]'.format(output_fname))
             except:
-                print(f'Conversion failed: ScanID:{str(scan_id)}, RecoID:{str(reco_id)}')
+                print('Conversion failed: ScanID:{}, RecoID:{}'.format(str(scan_id), str(reco_id)))
         else:
             for scan_id, recos in study._pvobj.avail_reco_id.items():
                 for reco_id in recos:
@@ -151,18 +151,18 @@ def main():
                         save_meta_files(study, args, scan_id, reco_id, output_fname)
                         print('NifTi file is genetared... [{}]'.format(output_fname))
                     except:
-                        print(f'Conversion failed: ScanID:{str(scan_id)}, RecoID:{str(reco_id)}')
+                        print('Conversion failed: ScanID:{}, RecoID:{}'.format(str(scan_id), str(reco_id)))
 
     elif args.function == 'tonii_all':
         from os.path import join as opj, isdir, isfile
 
         path = args.input
         slope, offset = set_rescale(args)
-        invalid_error_message = f'[Error] Invalid input path: {path}\n'
-        empty_folder = f'        The input path does not contain any raw data.'
-        wrong_target = f'        The input path indicates raw data itself. \n' \
-                       f'        You must input the parents folder instead of path of the raw data\n' \
-                       f'        If you want to convert single session raw data, use (tonii) instead.'
+        invalid_error_message = '[Error] Invalid input path: {}\n'.format(path)
+        empty_folder = '        The input path does not contain any raw data.'
+        wrong_target = '        The input path indicates raw data itself. \n' \
+                       '        You must input the parents folder instead of path of the raw data\n' \
+                       '        If you want to convert single session raw data, use (tonii) instead.'
 
         list_of_raw = sorted([d for d in os.listdir(path) if isdir(opj(path, d)) \
                               or (isfile(opj(path, d)) and (('zip' in d) or ('PvDataset' in d)))])
@@ -208,12 +208,12 @@ def main():
                                 study.save_as(scan_id, reco_id, output_fname, slope=slope, offset=offset)
                                 save_meta_files(study, args, scan_id, reco_id, output_fname)
                             except:
-                                print(f'Conversion failed: ScanID:{str(scan_id)}, RecoID:{str(reco_id)}')
-                    print(f'{raw} is converted...')
+                                print('Conversion failed: ScanID:{}, RecoID:{}'.format(str(scan_id), str(reco_id)))
+                    print('{} is converted...'.format(raw))
                 else:
-                    print(f'{raw} does not contains any scan data to convert...')
+                    print('{} does not contains any scan data to convert...'.format(raw))
             else:
-                print(f'{raw} is not PvDataset.')
+                print('{} is not PvDataset.'.format(raw))
 
     elif args.function == 'bids_helper':
         import pandas as pd
@@ -223,7 +223,7 @@ def main():
         make_json = args.json
         if not ds_output.endswith('.xlsx'):
             # to prevent pandas ValueError in case user does not provide valid file extension.
-            output = f'{ds_output}.xlsx'
+            output = '{}.xlsx'.format(ds_output)
         else:
             output = ds_output
 
@@ -282,9 +282,9 @@ def main():
 
         if make_json:
             fname = output[:-5]
-            json_fname = f'{fname}.json'
-            print(f'Creating JSON syntax template for parsing the BIDS required metadata '
-                  f'(BIDS v{_supporting_bids_ver}): {json_fname}')
+            json_fname = '{}.json'.format(fname)
[…]
