+++ paper/paper.md
@@ -21,72 +21,70 @@ affiliations:
    index: 1
  - name: Department of Neurology, University of North Carolina at Chapel Hill
    index: 2
- - name: Biomedical Research Imaging Center(BRIC), University of North Carolina at Chapel Hill 
+ - name: Biomedical Research Imaging Center(BRIC), University of North Carolina at Chapel Hill
    index: 3
 date: 22 April 2020
 bibliography: paper.bib
 ---
 
 # Summary
-Accessing and analyzing Magnetic Resonance Imaging (MRI) data required conversion from a hardware-specific format 
-to one compatible with available analysis software. The Digital Imaging and Communications in Medicine (DICOM) 
-is considered as an international standard for handling MRI data in the clinical field due to its flexibility 
-to cover the broad range of information from various imaging modalities and patient information. On the other hand, 
-the complex data structure and the inclusion of unnecessary metadata in DICOM decrease analysis efficiency 
-on various imaging research when the patient-related metadata is not essential for the study. 
-For this reason, image data formats with a more efficient structure capable of encompassing spatial and temporal 
-in a single file, such as NifTi-1, ANALYZE, or NRRD file formats, are becoming the new standard of choice 
-in many imaging research studies. Converting the raw data into DICOM format is still necessary for clinical research, 
-since it is important to preserve the patient-related metadata coupled with the image data. 
-However, the preclinical MRI research using laboratory animals or objects does not require the DICOM conversion step 
-and is rather inefficient. 
+BrkRaw is used for accessing and analyzing Magnetic Resonance Imaging (MRI) data required for conversion from a hardware-specific format
+to one that is compatible with available analysis software. The Digital Imaging and Communications in Medicine (DICOM)
+is considered an international standard for handling MRI data in the clinical field due to its ability
+to cover a broad range of information from various imaging modalities and patient information. On the other hand,
+when the patient-related metadata is not essential for the study, the complex data structure and the inclusion of unnecessary metadata in DICOM
+decreases the analysis efficiency on various imaging research. For this reason, image data formats with a more efficient structure that is capable of encompassing spatial and temporal analysis
+in a single file, such as NifTi-1, ANALYZE, or NRRD file formats, which are becoming the new standard of choice
+in many imaging research studies. Converting the raw data into DICOM format is still necessary for clinical research,
+since it is important to preserve the patient-related metadata coupled with the image data.
+However, the preclinical MRI research using laboratory animals or objects does not require the DICOM conversion step
+making it rather inefficient.
 
-Researchers have developed conversion software for Bruker Biospin MRI (the leading preclinical MRI system), 
-for converting raw data directly into a NIfTI or ANALYZE format 
-[@Brett:2002; @Ferraris:2017; @Chavarrias:2017; @Rorden:2018] for improving efficiency for data processing.
+Researchers have developed conversion software for Bruker Biospin MRI (the leading preclinical MRI system),
+for converting raw data directly into a NIfTI or ANALYZE format
+[@Brett:2002; @Ferraris:2017; @Chavarrias:2017; @Rorden:2018] in order to improve efficiency for data processing.
 
-Although the converters significantly improve analysis workflow by direct converting into software friendly file format,
-the software introduced so far has not provided the feature to access the data immediately without conversion.
+Although the converters significantly improve analysis workflow by directly converting data into a software friendly file format,
+the software introduced so far has not provided a feature that allows access to data immediately without conversion.
 Here we present the 'BrkRaw' python module, a more comprehensive preclinical tool for accessing and utilizing
-raw Bruker Biospin MRI data. The module has been built up upon robust low-level Python Application Programming Interface 
[…]
