### 2020-08-12 fd3520e
[update] CI workflow
