### 2020-08-21 7e4147b
[patch] loader.py - add 'FG_CARDIAC_MOVIE' to the available frame group
