### 2020-10-06 #2277 @richardjgowers
@labarba yeah I'm happy on using 2 reviewers here.  I was waiting on reviewer comments here: https://github.com/BrkRaw/bruker/issues/22 to get addressed by the authors (@dvm-shlee)
