### 2020-11-29 #2277 @labarba
Ping @openjournals/jose-eics  — I have retired from the JOSS board.
