### 2020-11-30 #2277 @arfon
@dvm-shlee - could you update us on your status here? Are you still interested in pursuing this publication in JOSS?
