### 2021-01-13 60a1ca1
[patch] loader.py - add 'FG_CARDIAC_MOVIE' to the available frame group
