+++ paper/paper.md
@@ -40,66 +40,52 @@ bibliography: paper.bib
 ---
 
 # Summary
+The Magnetic Resonance Imaging (MRI) data are acquired in a vendor-specific format which requiring 
+the conversion for the subsequent data analysis. The Digital Imaging and Communications in Medicine (DICOM) has been 
+used as an international standard for handling MRI data due to its medical application. On the other hand, 
+in the research field, the data formats with a simple structure that only composed of metadata that contains 
+information only for reconstructing binary data, such as NifTi-1, ANALYZE, or NRRD formats, have been adopted as 
+the new standard of choice because of their simplicity and efficiency. 
 
-Accessing and analyzing Magnetic Resonance Imaging (MRI) data required conversion from a hardware-specific format 
-to one compatible with available analysis software. The Digital Imaging and Communications in Medicine (DICOM) 
-is considered as an international standard for handling MRI data in the clinical field due to its flexibility 
-to cover the broad range of information from various imaging modalities and patient information. On the other hand, 
-the complex data structure and the inclusion of unnecessary metadata in DICOM decrease analysis efficiency 
-on various imaging research when the patient-related metadata is not essential for the study. 
-For this reason, image data formats with a more efficient structure capable of encompassing spatial and temporal 
-in a single file, such as NifTi-1, ANALYZE, or NRRD file formats, are becoming the new standard of choice 
-in many imaging research studies. Converting the raw data into DICOM format is still necessary for clinical research, 
-since it is important to preserve the patient-related metadata coupled with the image data. 
-However, the preclinical MRI research using laboratory animals or objects does not require the DICOM conversion step 
-and is rather inefficient. 
+Converting the raw data into DICOM format is still important to preserve the patient-related information coupled 
+with the images in the clinical field, however, in the preclinical setting, the data format does not need to be 
+internally complex so that the canonical way to initially convert data into DICOM and subsequent conversion makes 
+the overall analysis pipeline inefficient.
 
-Researchers have developed conversion software for Bruker Biospin MRI, 
-for converting raw data directly into a NIfTI or ANALYZE format 
-[@Brett:2002; @Ferraris:2017; @Chavarrias:2017; @Rorden:2018] for improving efficiency for data processing.
+Due to this reason, several researchers had developed tools to directly convert raw data to a NIfTI or ANALYZE format, 
+particularly for the preclinical MRI. [@Brett:2002; @Ferraris:2017; @Chavarrias:2017; @Rorden:2018]. These converters 
+significantly improved the efficiency of data analysis workflow for most applications. Still, it can be insufficient 
+for some circumstances when researchers develop a new imaging sequence that requiring new set parameters or 
+an additional layer of preprocessing for image reconstruction or when a quick quality assessment of raw data that 
+the providers' software does not support.
 
-Although the converters significantly improve analysis workflow by direct converting into software friendly file format,
-the software introduced so far has not provided the feature to access the data immediately without conversion.
-Here we present the 'BrkRaw' Python module, a more comprehensive preclinical tool for accessing and utilizing
-raw Bruker Biospin MRI data. The module has been built up upon robust low-level Python Application Programming Interface 
-(API), allowing direct raw data access without conversion to provide the advanced and easy-to-use features 
-for data analysis. 
+Here we present the 'BrkRaw' Python module, a more comprehensive tool to access and handle raw data for the preclinical 
+MRI scanner. Because the Bruker Biospin MRI scanner is the only ultra-high field preclinical MRI scanner currently 
+available in the market, we focussed on this specific vendor instead of developing a universal converter. 
+The BrkRaw module provides Application Programming Interface (API) and command-line tools to help integrate 
+into the data analysis pipeline more flexibly.
 
[…]
