### 2021-01-17 df6c9ae
[update] #13 paper.md: improve paper writing,
