### 2021-01-17 563cbb8
[patch] loader.py: change exception to warning for unexpected frame group.
