### 2021-03-22 d8343be
make bids_helper work on a single scan input foler as well as a parent folder
