### 2021-03-23 3f78d64
make bids_convert work on a single scan input foler as well as a parent folder
