+++ brkraw/scripts/brkraw.py
@@ -231,7 +231,7 @@ def main():
                    'task', 'acq', 'ce', 'rec', 'dir', 'run', 'modality', 'Start', 'End']
         df = pd.DataFrame(columns=Headers)
 
-        # if the path directly contains scan files
+        # if the path directly contains scan files for one participant
         if 'subject' in os.listdir(path):
             dNames = ['']
         else:         # old way, when you run against the parent folder (which contains one or more scan folder).
@@ -351,7 +351,14 @@ def main():
                 f.write('## How to cite?\n - https://doi.org/10.5281/zenodo.3818615\n')
 
         print('Inspect input BIDS datasheet...')
-        for dname in sorted(os.listdir(path)):
+
+        # if the path directly contains scan files for one participant
+        if 'subject' in os.listdir(path):
+            dNames = ['']
+        else:         # old way, when you run against the parent folder (which contains one or more scan folder).
+            dNames = sorted(os.listdir(path))
+
+        for dname in dNames:
             dpath = os.path.join(path, dname)
             try:
                 dset = BrukerLoader(dpath)
