### 2021-03-25 99ea8cf
init generateModalityAgnosticFiles
