### 2021-03-30 78de4a2
refactor createParticipantSessionFolder
