### 2021-03-30 f9aebe2
improve generateModalityAgnosticFiles to exit on exsiting files
