### 2021-03-31 6aeceee
update to createFolderTree
