### 2021-04-06 2e063aa
add clean Session ID
