+++ brkraw/scripts/brkraw.py
@@ -270,8 +270,6 @@ def main():
                                 if num_spack != 3:  # excluding localizer
                                     method = dset.get_method(scan_id).parameters['Method']
                                     if re.search('epi', method, re.IGNORECASE) and not re.search('dti', method, re.IGNORECASE):
-                                        # To use python debugger
-                                        import pdb; pdb.set_trace()
                                         #Why epi is function here? there should at lease a comment.
                                         datatype = 'func'
                                     elif re.search('dti', method, re.IGNORECASE):
