### 2021-05-17 34940d2
handle hyphen in partipcant ID and session ID
