### 2021-05-17 c82f540
add warning message for etc datatype
