### 2021-05-19 c243e40
fix RepetitionTime is mutually exclusive with VolumeTiming
