+++ brkraw/lib/loader.py
@@ -554,7 +554,8 @@ class BrukerLoader():
         if ('RepetitionTime' in json_obj.keys()) and ('VolumeTiming' in json_obj.keys()):
             if type(json_obj['RepetitionTime']) == int or type(json_obj['RepetitionTime']) == float:
                 del json_obj['VolumeTiming']
-                msg = "Both 'RepetitionTime' and 'VolumeTiming' exist in your json file, removed 'VolumeTiming' to make it valid for BIDS"
+                msg = "Both 'RepetitionTime' and 'VolumeTiming' exist in your .json file, removed 'VolumeTiming' to make it valid for BIDS.\
+                \n To use VolumeTiming, remove the RepetitionTime item but keep VolumeTiming from the .json file generated from bids_helper."
                 warnings.warn(msg)
 
         with open(os.path.join(dir, '{}.json'.format(filename)), 'w') as f:
