### 2021-05-24 8e71c00
fix issue related to gradient direction
