### 2021-05-24 4417296
add some minor comment for command line options
