+++ brkraw/scripts/brkraw.py
@@ -41,12 +41,15 @@ def main():
     bids_convert = subparsers.add_parser("bids_convert", help="Convert ALL raw Bruker data located "
                                                               "in the input directory based on the BIDS datasheet")
 
+    # Adding arguments for each parser
+    # gui
     gui.add_argument("-i", "--input", help=input_str, type=str, default=None)
     gui.add_argument("-o", "--output", help=output_dir_str, type=str, default=None)
     gui.add_argument("--ignore-slope", help='remove slope value from header', action='store_true')
     gui.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
     gui.add_argument("--ignore-rescale", help='remove slope and offset values from header', action='store_true')
 
+    # tonii
     nii.add_argument("input", help=input_str, type=str)
     nii.add_argument("-b", "--bids", help=bids_opt, action='store_true')
     nii.add_argument("-o", "--output", help=output_fnm_str, type=str, default=False)
@@ -58,6 +61,7 @@ def main():
     nii.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
     nii.add_argument("--ignore-rescale", help='remove slope and offset values from header', action='store_true')
 
+    # tonii_all
     niiall.add_argument("input", help=input_dir_str, type=str)
     niiall.add_argument("-o", "--output", help=output_dir_str, type=str)
     niiall.add_argument("-b", "--bids", help=bids_opt, action='store_true')
@@ -65,11 +69,13 @@ def main():
     niiall.add_argument("--ignore-offset", help='remove offset value from header', action='store_true')
     niiall.add_argument("--ignore-rescale", help='remove slope and offset values from header', action='store_true')
 
+    # bids_helper
     bids_helper.add_argument("input", help=input_dir_str, type=str)
     bids_helper.add_argument("output", help="output BIDS datasheet filename (.xlsx)", type=str)
     bids_helper.add_argument("-j", "--json", help="create JSON syntax template for "
                                                   "parsing metadata from the header", action='store_true')
 
+    # bids_convert
     bids_convert.add_argument("input", help=input_dir_str, type=str)
     bids_convert.add_argument("datasheet", help="input BIDS datahseet filename", type=str)
     bids_convert.add_argument("-j", "--json", help="input JSON syntax template filename", type=str, default=False)
