+++ brkraw/__init__.py
@@ -1,6 +1,6 @@
 from .lib import *
 
-__version__ = '0.3.6'
+__version__ = '0.3.7'
 __all__ = ['BrukerLoader', '__version__']
 
 
