### 2021-05-26 432cd7d
add MSME type and isolate assignDataType for bids helper
