### 2021-05-27 1a75efe
create new sub function for bids convert
