### 2021-06-07 0806ede
init session handle
