### 2021-06-08 3b53be3
make session default option
