+++ Dockerfile
@@ -1,3 +1,6 @@
+# How to use:
+# docker run -it --rm -v <your path to your data folder>:/data <your image>
+
 FROM python:3.7
 RUN mkdir /data
 RUN python -m pip install bruker==0.3.3
