### 2021-08-08 #2277 @arfon
@dvm-shlee - it has now been more than 6 months since we last heard from you. At this point I believe the best next step would be to close this review and encourage a resubmission in the future if you're able to improve the submission.

I'll set a reminder for two weeks to check in again but unless we hear back from you with a firm timeline in the next couple of weeks we will reject this submission.
