### 2021-08-18 9669c9d
Don't round header affine
We shouldn't impose any loss of precision, nifti headers already lose that due to float
