### 2021-08-21 45caa89
Update loader.py
paravision 360 compatibility
