### 2021-08-21 363fbee
Update loader.py
paravision 360 compatibility
