+++ brkraw/lib/loader.py
@@ -102,7 +102,13 @@ class BrukerLoader():
 
     @property
     def num_scans(self):
-        return len(self._pvobj._fid.keys())
+        # [20210820] Add-paravision 360 related.
+        # return len(self._pvobj._fid.keys())
+        len_scans = len(self._pvobj._fid.keys())
+        if len_scans > 0:
+            return len_scans
+        else:
+            return len(self._pvobj._2dseq.keys())
 
     @property
     def num_recos(self):
