### 2021-08-22 c79badc
Update loader.py
paravision 360 compatibility (Add VisuVersion 4)
